Change options from long to int.

+++ emc.h
@@ -16,24 +16,24 @@ typedef struct{
 }Setup;
 
   typedef struct{
-    long int model_side;
-    long int read_stride;
+    int model_side;
+    int read_stride;
     double wavelength;
     double pixel_size;
-    long int detector_size;
+    int detector_size;
     double detector_distance;
-    long int rotations_n;
+    int rotations_n;
     double sigma;
-    long int slice_chunk;
-    long int N_images;
-    long int max_iterations;
+    int slice_chunk;
+    int N_images;
+    int max_iterations;
     int blur_image;
     double blur_sigma;
     const char *mask_file;
     const char *image_prefix;
     int normalize_images;
     int known_intensity;
-    long int model_input;
+    int model_input;
     const char *model_file;
     int exclude_images;
     double exclude_ratio;
